Moving tests into single class for now and adding FeatureTogglingShouldntCauseError
